fix: modify the type of error text
