+++ .changeset/nervous-pets-sell.md
@@ -1,0 +1,5 @@
+---
+"@modern-js/module-tools": patch
+---
+
+Modify the type of error text and display log messages
