+++ packages/solutions/module-tools/src/tasks/generator-dts/index.ts
@@ -1,99 +1,22 @@
 import type { ChildProcess } from 'child_process';
 import * as path from 'path';
-import * as os from 'os';
 import { Import, fs } from '@modern-js/utils';
 import type { NormalizedConfig, CoreOptions } from '@modern-js/core';
-import type { ITsconfig } from '../types';
-
-const tsPathsTransform: typeof import('../utils/tspaths-transform') =
-  Import.lazy('../utils/tspaths-transform', require);
-const babel: typeof import('../utils/babel') = Import.lazy(
-  '../utils/babel',
-  require,
-);
-const constants: typeof import('./constants') = Import.lazy(
-  './constants',
-  require,
-);
+import type { ITsconfig } from '../../types';
 
 const core: typeof import('@modern-js/core') = Import.lazy(
   '@modern-js/core',
   require,
 );
-
 const execa: typeof import('execa') = Import.lazy('execa', require);
 const JSON5: typeof import('json5') = Import.lazy('json5', require);
 const argv: typeof import('process.argv').default = Import.lazy(
   'process.argv',
   require,
 );
-const deepMerge: typeof import('lodash.merge') = Import.lazy(
-  'lodash.merge',
-  require,
-);
-const glob: typeof import('glob') = Import.lazy('glob', require);
+const utils: typeof import('./utils') = Import.lazy('./utils', require);
 
 let removeTsconfigPath = '';
-
-const generatorTsConfig = (
-  projectTsconfig: ITsconfig,
-  {
-    appDirectory,
-    distDir,
-    sourceDir = 'src',
-  }: { appDirectory: string; distDir: string; sourceDir?: string },
-) => {
-  const tempPath = path.resolve(appDirectory, './node_modules');
-  const resolvePath = path.relative(tempPath, appDirectory);
-  // const rootDir = projectTsconfig.compilerOptions?.rootDir
-  //   ? path.join(resolvePath, projectTsconfig.compilerOptions?.rootDir)
-  //   : resolvePath;
-  // 目前限制 rootDir 的路径为 sourceDir
-  const rootDir = path.join(resolvePath, sourceDir);
-  const baseUrl = projectTsconfig.compilerOptions?.baseUrl
-    ? path.join(appDirectory, projectTsconfig.compilerOptions?.baseUrl)
-    : appDirectory;
-  // if include = ['src'], final include should be ['../src']
-  const include = projectTsconfig.include?.map(includePath =>
-    path.join(resolvePath, includePath),
-  );
-  const exclude = projectTsconfig.exclude?.map(excludePath =>
-    path.join(resolvePath, excludePath),
-  );
-
-  const resetConfig: ITsconfig = {
-    compilerOptions: {
-      rootDir,
-      baseUrl,
-      // Ensure that .d.ts files are created by tsc, but not .js files
-      declaration: true,
-      emitDeclarationOnly: true,
-      outDir: distDir,
-    },
-    include,
-    exclude,
-  };
-
-  const recommendOption = {
-    // Ensure that Babel can safely transpile files in the TypeScript project
-    compilerOptions: {
-      isolatedModules: true,
-    },
-  };
-
-  const tempTsconfigPath = path.join(tempPath, constants.tempTsconfigName);
-  fs.ensureFileSync(tempTsconfigPath);
-  fs.writeJSONSync(
-    tempTsconfigPath,
-    deepMerge(
-      recommendOption,
-      projectTsconfig,
-      // 此处是必须要覆盖用户默认配置
-      resetConfig,
-    ),
-  );
-  return tempTsconfigPath;
-};
 
 const getProjectTsconfig = (tsconfigPath: string | undefined): ITsconfig => {
   if (!tsconfigPath || !fs.existsSync(tsconfigPath)) {
@@ -152,7 +75,7 @@ const generatorDts = async (_: NormalizedConfig, config: IGeneratorConfig) => {
   } = config;
 
   const userTsconfig = getProjectTsconfig(tsconfigPath);
-  const willDeleteTsconfigPath = generatorTsConfig(userTsconfig, {
+  const willDeleteTsconfigPath = utils.generatorTsConfig(userTsconfig, {
     appDirectory,
     distDir,
     sourceDir: sourceDirName,
@@ -174,8 +97,8 @@ const generatorDts = async (_: NormalizedConfig, config: IGeneratorConfig) => {
     console.info('[Tsc Compiler]: Successfully');
   } catch (e) {
     if (!tsCheck) {
-      console.warn(
-        `There are some type issues, which can be checked by configuring 'source.enableTsChecker = true' ${os.EOL} in modern.config.js`,
+      console.info(
+        `There are some type warnings, which can be checked by configuring 'output.disableTsChecker = false'`,
       );
     } else {
       const isRecord = (input: any): input is Record<string, any> =>
@@ -187,38 +110,6 @@ const generatorDts = async (_: NormalizedConfig, config: IGeneratorConfig) => {
     }
   }
   fs.removeSync(willDeleteTsconfigPath);
-};
-
-const resolveAlias = (
-  modernConfig: NormalizedConfig,
-  config: IGeneratorConfig,
-  watchFilenames: string[] = [],
-) => {
-  const { output } = modernConfig;
-  const defaultPaths = { '@': ['./src'] };
-  const dtsDistPath = `${config.distDir}/**/*.d.ts`;
-  const dtsFilenames =
-    watchFilenames.length > 0
-      ? watchFilenames
-      : glob.sync(dtsDistPath, { absolute: true });
-  const alias = babel.getFinalAlias(modernConfig, {
-    appDirectory: config.projectData.appDirectory,
-    tsconfigPath:
-      config.tsconfigPath ||
-      path.join(config.projectData.appDirectory, './tsconfig.json'),
-    sourceAbsDir: config.distDir,
-  });
-  const mergedPaths = alias.isTsPath
-    ? alias.paths || {}
-    : { ...defaultPaths, ...(alias.paths || {}) };
-  const result = tsPathsTransform.transformDtsAlias({
-    filenames: dtsFilenames,
-    baseUrl: path.join(config.projectData.appDirectory, output.path || 'dist'),
-    paths: mergedPaths,
-  });
-  for (const r of result) {
-    fs.writeFileSync(r.path, r.content);
-  }
 };
 
 interface ITaskConfig {
@@ -256,8 +147,8 @@ const taskMain = async ({
     sourceDirName: config.sourceDirName,
   };
   await generatorDts(modernConfig, option);
-  // TODO: watch 模式下无法转换
-  resolveAlias(modernConfig, option);
+  // // TODO: watch 模式下无法转换
+  utils.resolveAlias(modernConfig, option);
 };
 
 (async () => {
