+++ packages/toolkit/compiler/babel/modern.config.js
@@ -1,7 +1,6 @@
 /** @type {import('@modern-js/module-tools').UserConfig} */
 module.exports = {
   output: {
-    // enableTsChecker: true,
     disableSourceMap: true,
   },
 };
